+++ src/linux/PeripheralBase.h
@@ -63,6 +63,7 @@ class PeripheralBase {
 
     bool _attempt_connect();
     bool _attempt_disconnect();
+    void _clear_characteristic_callbacks();
 
     std::shared_ptr<SimpleBluez::Characteristic> _get_characteristic(BluetoothUUID const& service_uuid,
                                                                      BluetoothUUID const& characteristic_uuid);