+++ src/linux/PeripheralBase.cpp
@@ -19,7 +19,7 @@ PeripheralBase::PeripheralBase(std::shared_ptr<SimpleBluez::Device> device,
 PeripheralBase::~PeripheralBase() {
     device_->clear_on_disconnected();
     device_->clear_on_services_resolved();
-    _clear_characteristic_callbacks();
+    _cleanup_characteristics();
 }
 
 std::string PeripheralBase::identifier() { return device_->name(); }
@@ -42,7 +42,7 @@ void PeripheralBase::connect() {
     // Set the on_disconnected callback once the connection attempts are finished, thus
     // preventing disconnection events that should not be seen by the user.
     device_->set_on_disconnected([this]() {
-        this->_clear_characteristic_callbacks();
+        this->_cleanup_characteristics();
         this->disconnection_cv_.notify_all();
 
         if (this->callback_on_disconnected_) {
@@ -223,7 +223,7 @@ void PeripheralBase::set_callback_on_disconnected(std::function<void()> on_disco
 
 // Private methods
 
-void PeripheralBase::_clear_characteristic_callbacks() {
+void PeripheralBase::_cleanup_characteristics() {
     // Get rid of all the callbacks to ensure that no invalid objects are being called.
 
     if (device_->has_battery_interface()) {
@@ -233,6 +233,10 @@ void PeripheralBase::_clear_characteristic_callbacks() {
     for (auto bluez_service : device_->services()) {
         for (auto bluez_characteristic : bluez_service->characteristics()) {
             bluez_characteristic->clear_on_value_changed();
+
+            if (bluez_characteristic->notifying()) {
+                bluez_characteristic->stop_notify();
+            }
         }
     }
 }
@@ -251,6 +255,8 @@ bool PeripheralBase::_attempt_connect() {
 }
 
 bool PeripheralBase::_attempt_disconnect() {
+    _cleanup_characteristics();
+
     device_->disconnect();
 
     // Wait for the disconnection to be confirmed.