+++ src/linux/Bluez.cpp
@@ -27,7 +27,11 @@ Bluez::~Bluez() {
 
 void Bluez::async_thread_function() {
     while (async_thread_active) {
-        bluez.run_async();
+        try {
+            bluez.run_async();
+        } catch (...) {
+            // TODO: Handle exception
+        }
         std::this_thread::sleep_for(std::chrono::microseconds(100));
     }
 }