+++ docs/changelog.rst
@@ -6,6 +6,22 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog`_, and this project adheres to
 `Semantic Versioning`_.
 
+[0.5.0] - 2022-XX-XX
+--------------------
+
+**Added**
+
+-
+
+**Changed**
+
+-  Clearer layout of examples. *(Thanks Yohannfra!)*
+
+**Fixed**
+
+-
+
+
 [0.4.0] - 2022-06-12
 --------------------